Move withContainer and withApiContainer to modules in middleware/
