+++ src/lib/middleware/withApiContainer.ts
@@ -1,0 +1,13 @@
+import { NextApiHandler, NextApiRequest, NextApiResponse } from "next";
+import { appContainer, IAppContainer } from "../appContainer";
+
+export type BeaconsApiRequest = NextApiRequest & {
+  container: Partial<IAppContainer>;
+};
+
+export const withApiContainer =
+  (callback: NextApiHandler): NextApiHandler =>
+  (req: BeaconsApiRequest, res: NextApiResponse) => {
+    req.container = req.container || appContainer;
+    return callback(req, res);
+  };
