+++ src/pages/api/registration/delete-use.ts
@@ -1,4 +1,7 @@
-import { BeaconsApiRequest, withApiContainer } from "../../../lib/container";
+import {
+  BeaconsApiRequest,
+  withApiContainer,
+} from "../../../lib/middleware/withApiContainer";
 
 export const handler = withApiContainer(async (req: BeaconsApiRequest, res) => {
   const { deleteCachedUse } = req.container;
