+++ src/pages/register-a-beacon/additional-beacon-use.tsx
@@ -5,12 +5,12 @@ import { BeaconUseSection } from "../../components/domain/BeaconUseSection";
 import { Grid } from "../../components/Grid";
 import { Layout } from "../../components/Layout";
 import { GovUKBody, PageHeading } from "../../components/Typography";
+import { showCookieBanner } from "../../lib/cookies";
+import { withCookieRedirect } from "../../lib/middleware";
 import {
   BeaconsGetServerSidePropsContext,
   withContainer,
-} from "../../lib/container";
-import { showCookieBanner } from "../../lib/cookies";
-import { withCookieRedirect } from "../../lib/middleware";
+} from "../../lib/middleware/withContainer";
 import { BeaconUse, IRegistration } from "../../lib/registration/types";
 import { retrieveUserFormSubmissionId } from "../../lib/retrieveUserFormSubmissionId";
 import { ActionURLs, PageURLs, queryParams } from "../../lib/urls";
