+++ src/pages/register-a-beacon/application-complete.tsx
@@ -6,12 +6,12 @@ import { Layout } from "../../components/Layout";
 import { Panel } from "../../components/Panel";
 import { GovUKBody, SectionHeading } from "../../components/Typography";
 import { WarningText } from "../../components/WarningText";
+import { verifyFormSubmissionCookieIsSet } from "../../lib/cookies";
+import { clearFormSubmissionCookie } from "../../lib/middleware";
 import {
   BeaconsGetServerSidePropsContext,
   withContainer,
-} from "../../lib/container";
-import { verifyFormSubmissionCookieIsSet } from "../../lib/cookies";
-import { clearFormSubmissionCookie } from "../../lib/middleware";
+} from "../../lib/middleware/withContainer";
 import { redirectUserTo } from "../../lib/redirectUserTo";
 import { retrieveUserFormSubmissionId } from "../../lib/retrieveUserFormSubmissionId";
 import { PageURLs } from "../../lib/urls";
