+++ src/useCases/getOrCreateAccountHolder.ts
@@ -1,6 +1,6 @@
 import { IAccountHolderDetails } from "../entities/accountHolderDetails";
 import { IAppContainer } from "../lib/appContainer";
-import { BeaconsGetServerSidePropsContext } from "../lib/container";
+import { BeaconsGetServerSidePropsContext } from "../lib/middleware/withContainer";
 
 export type GetOrCreateAccountHolderFn = (
   context: BeaconsGetServerSidePropsContext
